Django_api/requirements.txt
django        ==            5.*
djangorestframework         ==            3.16.*
pymongo     ==   4.13.2
pydantic
pydantic[email]        ==   2.7.4
pydantic-mongo          ==             3.1.0